+++ eslint.config.js
@@ -4,7 +4,7 @@ import tseslint from 'typescript-eslint';
 
 export default tseslint.config(
   {
-    ignores: ['dist', 'storybook-static', 'node_modules']
+    ignores: ['dist', 'storybook-static', '.playwright-browsers', 'node_modules']
   },
   js.configs.recommended,
   ...tseslint.configs.recommended,